+++ diff_test.go
@@ -2,9 +2,11 @@ package amt
 
 import (
 	"context"
+	"fmt"
 	"sort"
 	"strconv"
 	"testing"
+	"time"
 
 	cbor "github.com/ipfs/go-ipld-cbor"
 	"github.com/stretchr/testify/assert"
@@ -59,19 +61,12 @@ func TestSimpleEquals(t *testing.T) {
 	b, err := NewAMT(curBs)
 	assert.NoError(t, err)
 
-	_ = diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 0)
+	diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 0)
 
 	assertSet(t, a, 2, "foo")
 	assertSet(t, b, 2, "foo")
 
-	_ = diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 0)
-
-	_ = parallelDiffAndAssertLength(ctx, t, prevBs, curBs, a, b, 0)
-
-	assertSet(t, a, 2, "foo")
-	assertSet(t, b, 2, "foo")
-
-	_ = parallelDiffAndAssertLength(ctx, t, prevBs, curBs, a, b, 0)
+	diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 0)
 }
 
 func TestSimpleAdd(t *testing.T) {
@@ -96,8 +91,6 @@ func TestSimpleAdd(t *testing.T) {
 	assertGet(ctx, t, b, 5, "bar")
 	assertCount(t, b, 2)
 
-	cs := diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1)
-
 	ec := expectedChange{
 		Type:   Add,
 		Key:    5,
@@ -105,11 +98,7 @@ func TestSimpleAdd(t *testing.T) {
 		After:  "bar",
 	}
 
-	ec.assertExpectation(t, cs[0])
-
-	cs = diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1)
-
-	ec.assertExpectation(t, cs[0])
+	diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1, ec)
 }
 
 func TestDiffEmptyStateWithNonEmptyState(t *testing.T) {
@@ -129,8 +118,6 @@ func TestDiffEmptyStateWithNonEmptyState(t *testing.T) {
 		assertGet(ctx, t, prev, 2, "foo")
 		assertCount(t, prev, 1)
 
-		cs := diffAndAssertLength(ctx, t, prevBs, curBs, prev, cur, 1)
-
 		ec := expectedChange{
 			Type:   Remove,
 			Key:    2,
@@ -138,11 +125,7 @@ func TestDiffEmptyStateWithNonEmptyState(t *testing.T) {
 			After:  "",
 		}
 
-		ec.assertExpectation(t, cs[0])
-
-		cs = diffAndAssertLength(ctx, t, prevBs, curBs, prev, cur, 1)
-
-		ec.assertExpectation(t, cs[0])
+		diffAndAssertLength(ctx, t, prevBs, curBs, prev, cur, 1, ec)
 	})
 
 	t.Run("Added values", func(t *testing.T) {
@@ -161,8 +144,6 @@ func TestDiffEmptyStateWithNonEmptyState(t *testing.T) {
 		assertGet(ctx, t, cur, 2, "foo")
 		assertCount(t, cur, 1)
 
-		cs := diffAndAssertLength(ctx, t, prevBs, curBs, prev, cur, 1)
-
 		ec := expectedChange{
 			Type:   Add,
 			Key:    2,
@@ -170,12 +151,7 @@ func TestDiffEmptyStateWithNonEmptyState(t *testing.T) {
 			After:  "foo",
 		}
 
-		ec.assertExpectation(t, cs[0])
-
-		cs = parallelDiffAndAssertLength(ctx, t, prevBs, curBs, prev, cur, 1)
-
-		ec.assertExpectation(t, cs[0])
-
+		diffAndAssertLength(ctx, t, prevBs, curBs, prev, cur, 1, ec)
 	})
 }
 
@@ -200,8 +176,6 @@ func TestSimpleRemove(t *testing.T) {
 	assertSet(t, b, 2, "foo")
 	assertGet(ctx, t, b, 2, "foo")
 
-	cs := diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1)
-
 	ec := expectedChange{
 		Type:   Remove,
 		Key:    5,
@@ -209,11 +183,7 @@ func TestSimpleRemove(t *testing.T) {
 		After:  "",
 	}
 
-	ec.assertExpectation(t, cs[0])
-
-	cs = parallelDiffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1)
-
-	ec.assertExpectation(t, cs[0])
+	diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1, ec)
 }
 
 func TestSimpleModify(t *testing.T) {
@@ -230,8 +200,6 @@ func TestSimpleModify(t *testing.T) {
 	assertSet(t, a, 2, "foo")
 	assertSet(t, b, 2, "bar")
 
-	cs := diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1)
-
 	ec := expectedChange{
 		Type:   Modify,
 		Key:    2,
@@ -239,11 +207,7 @@ func TestSimpleModify(t *testing.T) {
 		After:  "bar",
 	}
 
-	ec.assertExpectation(t, cs[0])
-
-	cs = parallelDiffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1)
-
-	ec.assertExpectation(t, cs[0])
+	diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1, ec)
 }
 
 func TestLargeModify(t *testing.T) {
@@ -282,25 +246,7 @@ func TestLargeModify(t *testing.T) {
 		})
 	}
 
-	cs := diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 100)
-
-	sort.Slice(cs, func(i, j int) bool {
-		return cs[i].Key < cs[j].Key
-	})
-
-	for i := range cs {
-		ecs[i].assertExpectation(t, cs[i])
-	}
-
-	cs = parallelDiffAndAssertLength(ctx, t, prevBs, curBs, a, b, 100)
-
-	sort.Slice(cs, func(i, j int) bool {
-		return cs[i].Key < cs[j].Key
-	})
-
-	for i := range cs {
-		ecs[i].assertExpectation(t, cs[i])
-	}
+	diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 100, ecs...)
 }
 
 func TestLargeAdditions(t *testing.T) {
@@ -333,136 +279,104 @@ func TestLargeAdditions(t *testing.T) {
 		})
 	}
 
-	cs := diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 500)
-
-	sort.Slice(cs, func(i, j int) bool {
-		return cs[i].Key < cs[j].Key
-	})
-
-	for i := range cs {
-		ecs[i].assertExpectation(t, cs[i])
-	}
-
-	cs = parallelDiffAndAssertLength(ctx, t, prevBs, curBs, a, b, 500)
-
-	sort.Slice(cs, func(i, j int) bool {
-		return cs[i].Key < cs[j].Key
-	})
-
-	for i := range cs {
-		ecs[i].assertExpectation(t, cs[i])
-	}
+	diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 500, ecs...)
 }
 
 func TestBigDiff(t *testing.T) {
-	prevBs := cbor.NewCborStore(newMockBlocks())
-	curBs := cbor.NewCborStore(newMockBlocks())
-	ctx := context.Background()
+	for multiplier := 1; multiplier < 7; multiplier++ {
+		t.Run(fmt.Sprintf("Multiplier %d", multiplier), func(t *testing.T) {
 
-	a, err := NewAMT(prevBs)
-	assert.NoError(t, err)
+			prevBs := cbor.NewCborStore(newMockBlocks(WithGetDelay(time.Millisecond * 1)))
+			curBs := cbor.NewCborStore(newMockBlocks(WithGetDelay(time.Millisecond * 1)))
+			ctx := context.Background()
 
-	b, err := NewAMT(curBs)
-	assert.NoError(t, err)
-	multiplyer := 6
+			a, err := NewAMT(prevBs)
+			assert.NoError(t, err)
 
-	for i := 0; i < 100*multiplyer; i++ {
-		assertSet(t, a, uint64(i), "foo"+strconv.Itoa(i))
-	}
+			b, err := NewAMT(curBs)
+			assert.NoError(t, err)
 
-	ecs := make([]expectedChange, 0)
+			for i := 0; i < 100*multiplier; i++ {
+				assertSet(t, a, uint64(i), "foo"+strconv.Itoa(i))
+			}
 
-	// modify every other element, 50 modifies + 50 removes
-	for i := 0; i < 100*multiplyer; i += 2 {
-		assertSet(t, b, uint64(i), "bar"+strconv.Itoa(i))
+			ecs := make([]expectedChange, 0)
 
-		ecs = append(ecs, expectedChange{
-			Type:   Modify,
-			Key:    uint64(i),
-			Before: "foo" + strconv.Itoa(i),
-			After:  "bar" + strconv.Itoa(i),
+			// modify every other element, 50 modifies + 50 removes
+			for i := 0; i < 100*multiplier; i += 2 {
+				assertSet(t, b, uint64(i), "bar"+strconv.Itoa(i))
+
+				ecs = append(ecs, expectedChange{
+					Type:   Modify,
+					Key:    uint64(i),
+					Before: "foo" + strconv.Itoa(i),
+					After:  "bar" + strconv.Itoa(i),
+				})
+
+				ecs = append(ecs, expectedChange{
+					Type:   Remove,
+					Key:    uint64(i + 1),
+					Before: "foo" + strconv.Itoa(i+1),
+					After:  "",
+				})
+			}
+
+			// modify every element between 1000 and 1500, 500 modifies
+			for i := 1000 * multiplier; i < 1500*multiplier; i++ {
+				assertSet(t, a, uint64(i), "foo"+strconv.Itoa(i))
+				assertSet(t, b, uint64(i), "bar"+strconv.Itoa(i))
+
+				ecs = append(ecs, expectedChange{
+					Type:   Modify,
+					Key:    uint64(i),
+					Before: "foo" + strconv.Itoa(i),
+					After:  "bar" + strconv.Itoa(i),
+				})
+			}
+
+			// new additions, 500 additions
+			for i := 2000 * multiplier; i < 2500*multiplier; i++ {
+				assertSet(t, b, uint64(i), "bar"+strconv.Itoa(i))
+
+				ecs = append(ecs, expectedChange{
+					Type:   Add,
+					Key:    uint64(i),
+					Before: "",
+					After:  "bar" + strconv.Itoa(i),
+				})
+			}
+
+			// 10000-10249 is removed, 250 removals
+			for i := 10000 * multiplier; i < 10250*multiplier; i++ {
+				assertSet(t, a, uint64(i), "foo"+strconv.Itoa(i))
+
+				ecs = append(ecs, expectedChange{
+					Type:   Remove,
+					Key:    uint64(i),
+					Before: "foo" + strconv.Itoa(i),
+					After:  "",
+				})
+			}
+
+			// 10250-10500 is modified, 250 modifies
+			for i := 10250 * multiplier; i < 10500*multiplier; i++ {
+				assertSet(t, a, uint64(i), "foo"+strconv.Itoa(i))
+				assertSet(t, b, uint64(i), "bar"+strconv.Itoa(i))
+
+				ecs = append(ecs, expectedChange{
+					Type:   Modify,
+					Key:    uint64(i),
+					Before: "foo" + strconv.Itoa(i),
+					After:  "bar" + strconv.Itoa(i),
+				})
+			}
+
+			diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1600*multiplier, ecs...)
 		})
-
-		ecs = append(ecs, expectedChange{
-			Type:   Remove,
-			Key:    uint64(i + 1),
-			Before: "foo" + strconv.Itoa(i+1),
-			After:  "",
-		})
-	}
-
-	// modify every element between 1000 and 1500, 500 modifies
-	for i := 1000 * multiplyer; i < 1500*multiplyer; i++ {
-		assertSet(t, a, uint64(i), "foo"+strconv.Itoa(i))
-		assertSet(t, b, uint64(i), "bar"+strconv.Itoa(i))
-
-		ecs = append(ecs, expectedChange{
-			Type:   Modify,
-			Key:    uint64(i),
-			Before: "foo" + strconv.Itoa(i),
-			After:  "bar" + strconv.Itoa(i),
-		})
-	}
-
-	// new additions, 500 additions
-	for i := 2000 * multiplyer; i < 2500*multiplyer; i++ {
-		assertSet(t, b, uint64(i), "bar"+strconv.Itoa(i))
-
-		ecs = append(ecs, expectedChange{
-			Type:   Add,
-			Key:    uint64(i),
-			Before: "",
-			After:  "bar" + strconv.Itoa(i),
-		})
-	}
-
-	// 10000-10249 is removed, 250 removals
-	for i := 10000 * multiplyer; i < 10250*multiplyer; i++ {
-		assertSet(t, a, uint64(i), "foo"+strconv.Itoa(i))
-
-		ecs = append(ecs, expectedChange{
-			Type:   Remove,
-			Key:    uint64(i),
-			Before: "foo" + strconv.Itoa(i),
-			After:  "",
-		})
-	}
-
-	// 10250-10500 is modified, 250 modifies
-	for i := 10250 * multiplyer; i < 10500*multiplyer; i++ {
-		assertSet(t, a, uint64(i), "foo"+strconv.Itoa(i))
-		assertSet(t, b, uint64(i), "bar"+strconv.Itoa(i))
-
-		ecs = append(ecs, expectedChange{
-			Type:   Modify,
-			Key:    uint64(i),
-			Before: "foo" + strconv.Itoa(i),
-			After:  "bar" + strconv.Itoa(i),
-		})
-	}
-
-	cs := diffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1600*multiplyer)
-
-	sort.Slice(cs, func(i, j int) bool {
-		return cs[i].Key < cs[j].Key
-	})
-
-	for i := range cs {
-		ecs[i].assertExpectation(t, cs[i])
-	}
-
-	cs = parallelDiffAndAssertLength(ctx, t, prevBs, curBs, a, b, 1600*multiplyer)
-
-	sort.Slice(cs, func(i, j int) bool {
-		return cs[i].Key < cs[j].Key
-	})
-
-	for i := range cs {
-		ecs[i].assertExpectation(t, cs[i])
 	}
 }
 
-func diffAndAssertLength(ctx context.Context, t *testing.T, prevBs, curBs cbor.IpldStore, a *Root, b *Root, expectedLength int) []*Change {
+func diffAndAssertLength(ctx context.Context, t *testing.T, prevBs, curBs cbor.IpldStore, a *Root, b *Root, expectedLength int, ecs ...expectedChange) {
 	aCid, err := a.Flush(ctx)
 	if err != nil {
 		t.Fatal(err)
@@ -473,37 +387,50 @@ func diffAndAssertLength(ctx context.Context, t *testing.T, prevBs, curBs cbor.I
 		t.Fatal(err)
 	}
 
-	cs, err := Diff(ctx, prevBs, curBs, aCid, bCid)
-	if err != nil {
-		t.Fatalf("unexpected error from diff: %v", err)
-	}
+	var serial time.Duration
+	var parallel time.Duration
+	t.Run("assert serial diff", func(t *testing.T) {
+		start := time.Now()
+		cs, err := Diff(ctx, prevBs, curBs, aCid, bCid)
+		if err != nil {
+			t.Fatalf("unexpected error from diff: %v", err)
+		}
+		serial = time.Since(start)
 
-	if len(cs) != expectedLength {
-		t.Fatalf("got %d changes, wanted %d", len(cs), expectedLength)
-	}
+		if len(cs) != expectedLength {
+			t.Fatalf("got %d changes, wanted %d", len(cs), expectedLength)
+		}
 
-	return cs
-}
+		sort.Slice(cs, func(i, j int) bool {
+			return cs[i].Key < cs[j].Key
+		})
 
-func parallelDiffAndAssertLength(ctx context.Context, t *testing.T, prevBs, curBs cbor.IpldStore, a *Root, b *Root, expectedLength int) []*Change {
-	aCid, err := a.Flush(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
+		for i := range cs {
+			ecs[i].assertExpectation(t, cs[i])
+		}
+	})
 
-	bCid, err := b.Flush(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
+	t.Run("assert parallel diff", func(t *testing.T) {
+		start := time.Now()
+		cs, err := ParallelDiff(ctx, prevBs, curBs, aCid, bCid)
+		if err != nil {
+			t.Fatalf("unexpected error from diff: %v", err)
+		}
+		parallel = time.Since(start)
 
-	cs, err := ParallelDiff(ctx, prevBs, curBs, aCid, bCid)
-	if err != nil {
-		t.Fatalf("unexpected error from diff: %v", err)
-	}
+		if len(cs) != expectedLength {
+			t.Fatalf("got %d changes, wanted %d", len(cs), expectedLength)
+		}
 
-	if len(cs) != expectedLength {
-		t.Fatalf("got %d changes, wanted %d", len(cs), expectedLength)
-	}
+		sort.Slice(cs, func(i, j int) bool {
+			return cs[i].Key < cs[j].Key
+		})
 
-	return cs
+		for i := range cs {
+			ecs[i].assertExpectation(t, cs[i])
+		}
+
+	})
+	t.Logf("Serial Diff took   %s\tsize\t%d\tbitwidth %d", serial.Round(time.Millisecond), expectedLength, a.bitWidth)
+	t.Logf("Parallel Diff took %s\tsize\t%d\tbitwidth %d", parallel.Round(time.Millisecond), expectedLength, a.bitWidth)
 }
